TSK-1748: Review Findings
